Loads of exploration around being less exception oriented when something is wrong. Introduction of road network changes rejected event and rename to road network changes accepted event. Better split between validation responsibility and model responsibility (emerging). Various event property renames. Introduction of an explicit translator from the command change to the model's change.

+++ src/RoadRegistry/Commands/RoadSegmentHardeningProperties.cs
@@ -5,7 +5,7 @@ namespace RoadRegistry.Commands
     public class RoadSegmentHardeningProperties
     {
         public HardeningType Type { get; set; }
-        public decimal FromPosition { get; set; }
-        public decimal ToPosition { get; set; }
+        public double FromPosition { get; set; }
+        public double ToPosition { get; set; }
     }
 }
