+++ src/RoadRegistry/Commands/RoadSegmentWidthProperties.cs
@@ -3,7 +3,7 @@ namespace RoadRegistry.Commands
     public class RoadSegmentWidthProperties
     {
         public int Width { get; set; }
-        public decimal FromPosition { get; set; }
-        public decimal ToPosition { get; set; }
+        public double FromPosition { get; set; }
+        public double ToPosition { get; set; }
     }
 }
